+++ src/bundler/index.ts
@@ -11,7 +11,13 @@ function webpackConfig(
   isHook = false
 ): void {
   const scripts: any = BalmJS.config.scripts;
-  const defaultEntry = `./${BalmJS.config.roots.source}/${BalmJS.config.paths.source.js}/index.js`;
+  const defaultEntry =
+    './' +
+    path.join(
+      BalmJS.config.roots.source,
+      BalmJS.config.paths.source.js,
+      'index.js'
+    );
 
   const baseConfig: any = {
     entry: input ? getEntry(input, scripts) : defaultEntry,